Move Nimbus OAuth client into separate module
Closes #34487
