+++ test-framework/examples/tests/pom.xml
@@ -70,6 +70,10 @@
         </dependency>
         <dependency>
             <groupId>org.keycloak.test</groupId>
+            <artifactId>keycloak-test-framework-oauth-nimbus-poc</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.keycloak.test</groupId>
             <artifactId>keycloak-test-framework-ui</artifactId>
         </dependency>
         <dependency>
