+++ test-framework/examples/tests/src/test/java/org/keycloak/test/examples/EventsTest.java
@@ -5,12 +5,12 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.keycloak.events.EventType;
 import org.keycloak.test.framework.annotations.InjectEvents;
-import org.keycloak.test.framework.annotations.InjectOAuthClient;
+import org.keycloak.test.framework.oauth.nimbus.annotations.InjectOAuthClient;
 import org.keycloak.test.framework.ui.annotations.InjectPage;
 import org.keycloak.test.framework.ui.annotations.InjectWebDriver;
 import org.keycloak.test.framework.annotations.KeycloakIntegrationTest;
 import org.keycloak.test.framework.events.Events;
-import org.keycloak.test.framework.oauth.OAuthClient;
+import org.keycloak.test.framework.oauth.nimbus.OAuthClient;
 import org.keycloak.test.framework.ui.page.LoginPage;
 import org.keycloak.test.framework.server.KeycloakTestServerConfig;
 import org.openqa.selenium.WebDriver;
