+++ test-framework/examples/tests/src/test/java/org/keycloak/test/examples/OAuthClientTest.java
@@ -7,10 +7,10 @@ import com.nimbusds.oauth2.sdk.token.AccessToken;
 import jakarta.ws.rs.core.Response;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.keycloak.test.framework.annotations.InjectOAuthClient;
+import org.keycloak.test.framework.oauth.nimbus.annotations.InjectOAuthClient;
 import org.keycloak.test.framework.annotations.InjectUser;
 import org.keycloak.test.framework.annotations.KeycloakIntegrationTest;
-import org.keycloak.test.framework.oauth.OAuthClient;
+import org.keycloak.test.framework.oauth.nimbus.OAuthClient;
 import org.keycloak.test.framework.realm.ManagedUser;
 import org.keycloak.test.framework.realm.UserConfig;
 import org.keycloak.test.framework.realm.UserConfigBuilder;
