+++ test-framework/oauth-nimbus-poc/src/main/resources/META-INF/services/org.keycloak.test.framework.TestFrameworkExtension
@@ -1,0 +1,1 @@
+org.keycloak.test.framework.oauth.nimbus.NimbusTestFrameworkExtension
