+++ test-framework/pom.xml
@@ -40,6 +40,7 @@
         <module>db-mysql</module>
         <module>db-oracle</module>
         <module>db-postgres</module>
+        <module>oauth-nimbus-poc</module>
         <module>ui</module>
         <module>examples</module>
     </modules>
